+++ lib/resque_aps.rb
@@ -10,6 +10,7 @@ require 'resque/plugins/aps/application'
 require 'resque/plugins/aps/notification'
 require 'resque/plugins/aps/feedback'
 require 'resque/plugins/aps/unknown_attribute_error'
+require 'resque/plugins/aps/c2dm_socket'
 
 Resque.extend Resque::Plugins::Aps
 Resque::Server.class_eval do